+++ plugin-templates/typescript-template/commands/command.ts
@@ -1,9 +1,11 @@
 //! THE COMMAND, STORED IN A VARIABLE
 
-const command = {
+import { CommandInterface } from "scaped/types";
+
+const command: CommandInterface = {
     // Even if you don't want to specify an alias,
     // the strings need to be in an array, or else an error will occur in our plugin handler.
-    name: ["command", "myCommandAlias"], 
+    name: ["hello", "hey"], 
     
     // You don't necessarily need to use args at all if the command is simply one word.
     // Any validation, such as if the args you need exist or not, the validation is up to you.
@@ -14,11 +16,11 @@ const command = {
     // The function has to take in the arguements from the CLI, even if you don't need them.
     // If you don't need them, you can put it in the parameters and then ignore it.
     execute: async (args) => {
-        console.log("my command be like")
+        console.log("Hello World!")
         // Example of validating and handling args:
 
         if (!args.myArg) {
-            console.log("Hey! You didn't define myArg!");
+            console.log("Hey! You didn't define --myArg!");
             return;
         } else if (typeof args.myArg !== "boolean") {
             console.log("Hey! myArg should be a boolean which is true!");