+++ src/types.ts
@@ -5,6 +5,14 @@ interface CommandInterface {
     execute: (args: yargsParser.Arguments) => any
 }
 
+interface PluginConfig {
+    name: string;
+    version: string;
+    author: string;
+    lang: "javascript" | "typescript"
+}
+
 export {
-    CommandInterface
+    CommandInterface,
+    PluginConfig
 }