+++ types.d.ts
@@ -2,3 +2,4 @@
 // not done
 
 export type { CommandInterface } from "./src/types";
+export type { PluginConfig } from "./src/types";